use tanstack. Migrate code
