+++ package.json
@@ -14,6 +14,7 @@
     "@radix-ui/react-slot": "^1.1.0",
     "@remoteoss/json-schema-form": "^0.11.2-beta.0",
     "@rjsf/core": "^5.20.0",
+    "@tanstack/react-query": "^5.59.0",
     "axios": "^1.7.4",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
